[ElementTiming] Mimic FCP++ rect computations
This CL changes ElementTiming rect computation to use a computation
similar to that used by FCP++, but relative to the owning frame.

The tests show that the new computation seems more accurate as they
uncovered a problem with the SVG test: default SVG size is 300x150 so
it was clipping the <image> element inside.

A test for rect coordinates relative to iframe is added.

Bug: 879270, 943138
Change-Id: Ib8fa266bda76843fbf7ca07f82ac5bb245d52c4d