Barcode Detection Mac: Accept barcode format hint in detector constructor.
A user may specify barcode formats to which to restrict detection, which
are converted internally to all known related variants.

For Mac, this is done by setting the "symbologies" property of a Vision
API request. This request created and performs detection in util functions,
and otherwise is handled in wrapped form, so introduce two callbacks:
* SetupCallback in detection_utils_mac.h, which operates on a request
* SetupOptionsCallback in barcode_detection_impl_mac_vision.h, which
  becomes a SetupCallback by specifying how a request processes
  BarcodeDetectorOptions.

Change-Id: Ia12d250f9a753187c47c85b0fbc0765840e865ab