Implement `Sec-Fetch-Site: 'none'` for user-initiated navigations.
https://mikewest.github.io/sec-metadata/#sec-fetch-site-header defines
a `none` value for `Sec-Fetch-Site`, which should be delivered along with
navigational requests which were initiated directly from the browser, and
not from a user's interaction with websites.

This patch implements this new value in
`NavigationRequest::AddAdditionalRequestHeaders`, and adds tests for some
of the core use cases we expect servers to care about when examining
this header: address bar navigations, and bookmark navigations.

We'll add additional test cases in future patches as we work out the right
answers to the questions raised in [1] and [2].

[1]: https://mikewest.github.io/sec-metadata/#directly-user-initiated
[2]: https://docs.google.com/document/d/1YdakWfCR29pVNHURgceKrM_zjGN61xdnCXdZIGpQzxQ

Bug: 868286
Change-Id: I4866daba27a8763966cf3b213d0c0e916c0db787